Copied changes from WIP branch + copied the latest ConfigureConnection struct to RealWorkerConnection.cpp
